Create project_charter.md
This task was assigned to me by someone in the group.  I completed this before getting access to the repo.  I was invited as a collaborator, but I was never given the link so I did not have access.  I am committing this to show that I did do work, but the group completed things without me.